requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Pydantic
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Database (async SQLAlchemy + SQLite + Postgres drivers)
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
asyncpg==0.29.0          # for Postgres in production

# Auth
bcrypt==4.2.0
python-jose[cryptography]==3.3.0

# Utilities
python-multipart==0.0.12
aiofiles==24.1.0
httpx==0.27.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
